Move resover stuff to tcp where it is useful
